Slackbot whitelisting improvements
This PR includes two whitelisting improvements.

The first is adding support for a `never` expiration value. This will
set the TTL to ten years from now.

The second is that when an IP is whitelisted, the slackbot will iterate
through all provided iprepd urls and send DELETE requests.

As well, this PR switches slackbot-background to using sops and includes
a new `Configuration` struct that abstracts this (so that all of the
other cloudfunctions can start using it).

This was done in this PR as trying to shove multiple sets of iprepd
urls/api key combinations into a single environment variable was getting
finicky.

Bug: https://bugzilla.mozilla.org/show_bug.cgi?id=1557140