Auth0Pull cloud function
Auth0Pull is essentially duopull, but for auth0 log messages.

Auth0 logs are pulled starting at a set "last log id", then the last log
id is stored for the next run. Logs are then written to Stackdriver.

This will be deployed with a cloud scheduler job that will trigger it
every X minutes.